refactor: Extract PathPart to its own module
And be more explicit about the privacy of the inner String
