+++ object_store/src/path.rs
@@ -1,7 +1,6 @@
 //! This module contains code for abstracting object locations that work
 //! across different backing implementations and platforms.
 
-use percent_encoding::{percent_decode_str, percent_encode, AsciiSet, CONTROLS};
 use std::{mem, path::PathBuf};
 
 /// Paths that came from or are to be used in cloud-based object storage
@@ -13,6 +12,9 @@ pub mod file;
 /// Maximally processed storage-independent paths.
 pub mod parsed;
 use parsed::DirsAndFileName;
+
+mod parts;
+use parts::PathPart;
 
 /// Universal interface for handling paths and locations for objects and
 /// directories in the object store.
@@ -214,102 +216,10 @@ impl PartialEq for PathRepresentation {
 
 /// The delimiter to separate object namespaces, creating a directory structure.
 pub const DELIMITER: &str = "/";
-// percent_encode's API needs this as a byte
-const DELIMITER_BYTE: u8 = DELIMITER.as_bytes()[0];
-
-/// The PathPart type exists to validate the directory/file names that form part
-/// of a path.
-///
-/// A PathPart instance is guaranteed to contain no `/` characters as it can
-/// only be constructed by going through the `try_from` impl.
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Default)]
-pub struct PathPart(String);
-
-/// Characters we want to encode.
-const INVALID: &AsciiSet = &CONTROLS
-    // The delimiter we are reserving for internal hierarchy
-    .add(DELIMITER_BYTE)
-    // Characters AWS recommends avoiding for object keys
-    // https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingMetadata.html
-    .add(b'\\')
-    .add(b'{')
-    // TODO: Non-printable ASCII characters (128–255 decimal characters)
-    .add(b'^')
-    .add(b'}')
-    .add(b'%')
-    .add(b'`')
-    .add(b']')
-    .add(b'"') // " <-- my editor is confused about double quotes within single quotes
-    .add(b'>')
-    .add(b'[')
-    .add(b'~')
-    .add(b'<')
-    .add(b'#')
-    .add(b'|')
-    // Characters Google Cloud Storage recommends avoiding for object names
-    // https://cloud.google.com/storage/docs/naming-objects
-    .add(b'\r')
-    .add(b'\n')
-    .add(b'*')
-    .add(b'?');
-
-impl From<&str> for PathPart {
-    fn from(v: &str) -> Self {
-        match v {
-            // We don't want to encode `.` generally, but we do want to disallow parts of paths
-            // to be equal to `.` or `..` to prevent file system traversal shenanigans.
-            "." => Self(String::from("%2E")),
-            ".." => Self(String::from("%2E%2E")),
-            other => Self(percent_encode(other.as_bytes(), INVALID).to_string()),
-        }
-    }
-}
-
-impl std::fmt::Display for PathPart {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        percent_decode_str(&self.0)
-            .decode_utf8()
-            .expect("Valid UTF-8 that came from String")
-            .fmt(f)
-    }
-}
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn path_part_delimiter_gets_encoded() {
-        let part: PathPart = "foo/bar".into();
-        assert_eq!(part, PathPart(String::from("foo%2Fbar")));
-    }
-
-    #[test]
-    fn path_part_gets_decoded_for_display() {
-        let part: PathPart = "foo/bar".into();
-        assert_eq!(part.to_string(), "foo/bar");
-    }
-
-    #[test]
-    fn path_part_given_already_encoded_string() {
-        let part: PathPart = "foo%2Fbar".into();
-        assert_eq!(part, PathPart(String::from("foo%252Fbar")));
-        assert_eq!(part.to_string(), "foo%2Fbar");
-    }
-
-    #[test]
-    fn path_part_cant_be_one_dot() {
-        let part: PathPart = ".".into();
-        assert_eq!(part, PathPart(String::from("%2E")));
-        assert_eq!(part.to_string(), ".");
-    }
-
-    #[test]
-    fn path_part_cant_be_two_dots() {
-        let part: PathPart = "..".into();
-        assert_eq!(part, PathPart(String::from("%2E%2E")));
-        assert_eq!(part.to_string(), "..");
-    }
 
     // Invariants to maintain/document/test:
     //
@@ -507,10 +417,10 @@ mod tests {
         let path_buf_parts: DirsAndFileName = path_buf.into();
 
         assert!(cloud_parts.directories.is_empty());
-        assert_eq!(cloud_parts.file_name.unwrap().0, "blah.json");
+        assert_eq!(cloud_parts.file_name.unwrap().encoded(), "blah.json");
 
         assert!(path_buf_parts.directories.is_empty());
-        assert_eq!(path_buf_parts.file_name.unwrap().0, "blah.json");
+        assert_eq!(path_buf_parts.file_name.unwrap().encoded(), "blah.json");
 
         // empty
         let cloud = PathRepresentation::RawCloud("".into());
